Tarek

+++ app/Http/Controllers/Admin/OneAboutController.php
@@ -43,11 +43,13 @@ class OneAboutController extends Controller
         if(isset($data['first_photo'])) {
             saveFiles($data['first_photo'], $oneAbout, 'oneAbout', "first_photo",'create');
         }
-        foreach ($data['details'] as $detail) {
-            if(isset($detail['image']) && is_file($detail['image'])){
-                $detail['image'] = saveFile($detail['image'], 'oneAboutDetails');
+        if (!empty($data['details'])) {
+            foreach ($data['details'] as $detail) {
+                if(isset($detail['image']) && is_file($detail['image'])){
+                    $detail['image'] = saveFile($detail['image'], 'oneAboutDetails');
+                }
+                AboutDetail::create(array_merge(['one_about_id' => $oneAbout->id],$detail));
             }
-            $d = AboutDetail::create(array_merge(['one_about_id' => $oneAbout->id],$detail));
         }
         return responseJson([],'Added Successfully',200);
     }
@@ -60,20 +62,23 @@ class OneAboutController extends Controller
         }
 
         $oneAbout->update(Arr::except($data,['details','first_photo']));
-        // Delete old details
-        foreach ($oneAbout->details as $detail) {
-            $detail->delete();
-        }
-        // Save new details
-        foreach ($data['details'] as $detail) {
-            if(isset($detail['image']) && is_file($detail['image'])){
-                $detail['image'] = saveFile($detail['image'], 'oneAboutDetails');
-            } elseif (isset($detail['old_image'])) {
-                $detail['image'] = $detail['old_image'];
-                unset($detail['old_image']);
+
+        if (array_key_exists('details', $data)) {
+            foreach ($oneAbout->details as $detail) {
+                $detail->delete();
             }
-            $d = AboutDetail::create(array_merge(['one_about_id' => $oneAbout->id],$detail));
+
+            foreach ($data['details'] ?? [] as $detail) {
+                if(isset($detail['image']) && is_file($detail['image'])){
+                    $detail['image'] = saveFile($detail['image'], 'oneAboutDetails');
+                } elseif (isset($detail['old_image'])) {
+                    $detail['image'] = $detail['old_image'];
+                    unset($detail['old_image']);
+                }
+                AboutDetail::create(array_merge(['one_about_id' => $oneAbout->id],$detail));
+            }
         }
+
         return responseJson([],'Updated Successfully',200);
     }
